exception errors, buy button, jsonb objects

+++ src/main/java/com/rudii/Stock/model/Positions.java
@@ -4,13 +4,13 @@ public class Positions {
 
     private String symbol;
     private int shares;
-    private long price_per_share;
+    private int price_per_share;
 
     public Positions() {
 
     }
 
-    public Positions(String symbol, int shares, long price_per_share) {
+    public Positions(String symbol, int shares, int price_per_share) {
         this.symbol = symbol;
         this.shares = shares;
         this.price_per_share = price_per_share;
@@ -32,11 +32,11 @@ public class Positions {
         this.shares = shares;
     }
 
-    public long getPrice_per_share() {
+    public int getPrice_per_share() {
         return price_per_share;
     }
 
-    public void setPrice_per_share(long price_per_share) {
+    public void setPrice_per_share(int price_per_share) {
         this.price_per_share = price_per_share;
     }
 }
